Moved VariableNotUsedInspectionTests for QuickFixes into relevant files under RD.QuickFixes
Additionally cleared up some confused test names
