+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -1058,6 +1058,33 @@ End Sub";
             Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void UnusedVariable_IgnoreQuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo()
+Dim var1 As String
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+'@Ignore VariableNotUsed
+Dim var1 As String
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new VariableNotUsedInspection(state);
+            new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspection.GetInspectionResults().First());
+
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+
+
     }
 }
