+++ RubberduckTests/QuickFixes/RemoveUnusedDeclarationQuickFixTests.cs
@@ -2,7 +2,6 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Rubberduck.Inspections.Concrete;
 using Rubberduck.Inspections.QuickFixes;
-using Rubberduck.Parsing.Inspections.Resources;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 using RubberduckTests.Mocks;
 
@@ -117,6 +116,31 @@ End Sub";
             var rewriter = state.GetRewriter(component);
             Assert.AreEqual(expectedCode, rewriter.GetText());
         }
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void UnassignedVariable_QuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo()
+Dim var1 As String
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new VariableNotUsedInspection(state);
+            new RemoveUnusedDeclarationQuickFix(state).Fix(inspection.GetInspectionResults().First());
+
+            var rewriter = state.GetRewriter(component);
+            Assert.AreEqual(expectedCode, rewriter.GetText());
+        }
+
 
     }
 }
